+++ TargetLibraries/Generic/src/RQGELU_s8.c
@@ -28,8 +28,8 @@
 
 #include "DeeployBasicMath.h"
 
-void RQGELU_s8_s8(int8_t *data_in, int8_t *data_out, int32_t dataSize, int8_t b,
-                  int16_t one, int32_t input_offset, int32_t output_offset,
+void RQGELU_s8_s8(int8_t *data_in, int8_t *data_out, int32_t dataSize, int32_t b,
+                  int32_t one, int32_t input_offset, int32_t output_offset,
                   int32_t *mul, int32_t *add, int32_t *shift) {
 
   int32_t sign, x, x_abs, q;
